t.rast.algebra tests: split into two files
Since the `testsuite/test_raster_algebra.py` comes with many tests which overall may take more than 300s of runtime, this PR separates out the `test_temporal*()` tests into an own file. Thus, in the ideal case, a CI timeout is no longer reached in this case.

Partially addresses #2185
