+++ temporal/t.rast.algebra/testsuite/test_raster_algebra.py
@@ -119,33 +119,6 @@ class TestTRastAlgebra(TestCase):
         cls.runModule("t.unregister", maps="singletmap", quiet=True)
         cls.del_temp_region()
 
-    def test_temporal_conditional_time_dimension_bug(self):
-        """Testing the conditional time dimension bug, that uses the time
-        dimension of the conditional statement instead the time dimension
-        of the then/else statement."""
-        self.assertModule(
-            "t.rast.algebra",
-            expression="R = if({contains}, B == 5, " "A - 1,  A + 1)",
-            basename="r",
-            flags="d",
-        )
-        self.assertModule(
-            "t.rast.algebra",
-            expression="R = if({contains}, B == 5, " "A - 1,  A + 1)",
-            basename="r",
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 4)
-        self.assertEqual(D.metadata.get_min_min(), 0)  # 1 - 1
-        self.assertEqual(D.metadata.get_max_max(), 5)  # 4 + 1
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
     def test_simple_arith_hash_1(self):
         """Simple arithmetic test including the hash operator"""
 
@@ -357,159 +330,6 @@ class TestTRastAlgebra(TestCase):
         self.assertEqual(start, datetime.datetime(2001, 1, 1))
         self.assertEqual(end, datetime.datetime(2001, 1, 5))
 
-    def test_temporal_intersection_1(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {+,equal,i} B", basename="r"
-        )
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 0)
-
-    def test_temporal_intersection_2(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {+,during,i} B", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 4)
-        self.assertEqual(D.metadata.get_min_min(), 6)  # 1 + 5
-        self.assertEqual(D.metadata.get_max_max(), 10)  # 4 + 6
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-
-    def test_temporal_intersection_3(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {+,starts,i} B", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 6)  # 1 + 5
-        self.assertEqual(D.metadata.get_max_max(), 9)  # 3 + 6
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
-
-    def test_temporal_intersection_4(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {+,finishes,intersect} B", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 7)  # 2 + 5
-        self.assertEqual(D.metadata.get_max_max(), 10)  # 4 + 6
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-
-    def test_temporal_intersection_5(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {+,starts|finishes,i} B", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 4)
-        self.assertEqual(D.metadata.get_min_min(), 6)  # 1 + 5
-        self.assertEqual(D.metadata.get_max_max(), 10)  # 4 + 6
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-
-    def test_temporal_intersection_6(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = B {+,overlaps,u} C", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 1)
-        self.assertEqual(D.metadata.get_min_min(), 12)  # 5 + 7
-        self.assertEqual(D.metadata.get_max_max(), 12)  # 5 + 7
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
-
-    def test_temporal_intersection_7(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = B {+,overlapped,u} C", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 1)
-        self.assertEqual(D.metadata.get_min_min(), 13)  # 6 + 7
-        self.assertEqual(D.metadata.get_max_max(), 13)  # 6 + 7
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-
-    def test_temporal_intersection_8(self):
-        """Simple temporal intersection test"""
-
-        self.assertModule(
-            "t.rast.algebra",
-            expression='R = A {+,during,l} buff_t(C, "1 day") ',
-            basename="r",
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 4)
-        self.assertEqual(D.metadata.get_min_min(), 8)  # 1 + 7  a1 + c1
-        self.assertEqual(D.metadata.get_max_max(), 11)  # 4 + 7  a4 + c1
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-
-    def test_temporal_neighbors_1(self):
-        """Simple temporal neighborhood computation test"""
-
-        self.assertModule("t.rast.algebra", expression="R = A[-1] + A[1]", basename="r")
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 4)  # 1 + 3
-        self.assertEqual(D.metadata.get_max_max(), 6)  # 2 + 4
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
-
-    def test_temporal_neighbors_2(self):
-        """Simple temporal neighborhood computation test"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A[0,0,-1] + A[0,0,1]", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 4)  # 1 + 3
-        self.assertEqual(D.metadata.get_max_max(), 6)  # 2 + 4
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
 
     def test_tmap_function1(self):
         """Testing the tmap function."""
@@ -587,186 +407,6 @@ class TestTRastAlgebra(TestCase):
         self.assertEqual(D.check_temporal_topology(), True)
         self.assertEqual(D.get_granularity(), "1 day")
 
-    def test_temporal_select(self):
-        """Testing the temporal select operator."""
-
-        self.assertModule("t.rast.algebra", expression="R = A : A", basename="r")
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 4)
-        self.assertEqual(D.metadata.get_min_min(), 1)
-        self.assertEqual(D.metadata.get_max_max(), 4)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
-    def test_temporal_select(self):
-        """Testing the temporal select operator."""
-
-        self.assertModule("t.rast.algebra", expression="R = A : D", basename="r")
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 3)
-        self.assertEqual(D.metadata.get_max_max(), 4)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 3))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
-    def test_temporal_select_operators1(self):
-        """Testing the temporal select operator. Including temporal relations."""
-
-        self.assertModule("t.rast.algebra", expression="R = A : D", basename="r")
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 3)
-        self.assertEqual(D.metadata.get_max_max(), 4)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 3))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
-    def test_temporal_select_operators2(self):
-        """Testing the temporal select operator. Including temporal relations."""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {!:,during} C", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 1)
-        self.assertEqual(D.metadata.get_max_max(), 4)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
-    def test_temporal_select_operators3(self):
-        """Testing the temporal select operator. Including temporal relations and
-        different temporal operators (lr|+&)"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {:,during,d} B", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 4)
-        self.assertEqual(D.metadata.get_min_min(), 1)
-        self.assertEqual(D.metadata.get_max_max(), 4)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 1))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), False)
-        self.assertEqual(D.get_granularity(), "2 days")
-
-    def test_temporal_select_operators4(self):
-        """Testing the temporal select operator. Including temporal relations and
-        different temporal operators (lr|+&)"""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = A {:,equal|during,r} C", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        maplist = D.get_registered_maps_as_objects()
-        for map_i in maplist:
-            start_map, end_map = map_i.get_absolute_time()
-            self.assertEqual(start_map, datetime.datetime(2001, 1, 2))
-            self.assertEqual(end_map, datetime.datetime(2001, 1, 4))
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 2)
-        self.assertEqual(D.metadata.get_max_max(), 3)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
-        self.assertEqual(D.check_temporal_topology(), False)
-        self.assertEqual(D.get_granularity(), "2 days")
-
-    def test_temporal_hash_operator1(self):
-        """Testing the temporal hash operator in the raster algebra."""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = if(A # D == 1, A)", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 3)
-        self.assertEqual(D.metadata.get_max_max(), 4)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 3))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
-    def test_temporal_hash_operator2(self):
-        """Testing the temporal hash operator in the raster algebra."""
-
-        self.assertModule("t.rast.algebra", expression="R = A # D", basename="r")
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 2)
-        self.assertEqual(D.metadata.get_min_min(), 1)
-        self.assertEqual(D.metadata.get_max_max(), 1)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 3))
-        self.assertEqual(end, datetime.datetime(2001, 1, 5))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "1 day")
-
-    def test_temporal_hash_operator3(self):
-        """Testing the temporal hash operator in the raster algebra."""
-
-        self.assertModule(
-            "t.rast.algebra", expression="R = C {#,contains} A", basename="r"
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 1)
-        self.assertEqual(D.metadata.get_min_min(), 2)
-        self.assertEqual(D.metadata.get_max_max(), 2)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "2 days")
-
-    def test_temporal_hash_operator4(self):
-        """Testing the temporal hash operator in the raster algebra."""
-
-        self.assertModule(
-            "t.rast.algebra",
-            expression="R = if({contains},A # D == 1, C {#,contains} A)",
-            basename="r",
-        )
-
-        D = tgis.open_old_stds("R", type="strds")
-
-        self.assertEqual(D.metadata.get_number_of_maps(), 1)
-        self.assertEqual(D.metadata.get_min_min(), 2)
-        self.assertEqual(D.metadata.get_max_max(), 2)
-        start, end = D.get_absolute_time()
-        self.assertEqual(start, datetime.datetime(2001, 1, 2))
-        self.assertEqual(end, datetime.datetime(2001, 1, 4))
-        self.assertEqual(D.check_temporal_topology(), True)
-        self.assertEqual(D.get_granularity(), "2 days")
 
     def test_raster_arithmetic_relation_1(self):
         """Arithmetic test with temporal intersection"""
